fix : 컨트롤러 구조 분리
